services/detector/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
numpy>=1.24.0
pydantic==2.5.2
httpx==0.25.2
opencv-python-headless>=4.8.0
ultralytics>=8.0.0
Pillow>=10.0.0